changed the type of the data from float64 to int32 as it should be -- initial data is int16

+++ Cosmicremoval_temporal_ndbins.py
@@ -162,7 +162,7 @@ class Cosmicremoval_class:
                 else:
                     temp = temp2
                 data = hdul[detector].data
-                images.append(np.double(data[0, :, :, 0]))
+                images.append(np.double(data[0, :, :, 0]).astype(int))
                 dsun.append(dist_SUN)
                 temps.append(temp)
                 dates.append(date)
